requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.8.0
django-jazzmin==2.4.8
django-summernote==0.8.11.6
djangorestframework==3.12.4
django-filter==22.1
Pillow==8.3.1
python-decouple==3.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
django-debug-toolbar==3.5.0